requirements.txt
APScheduler==3.6.3
Cython==0.29.21
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
dm-tree==0.1.5
ecdsa==0.16.1
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gin-config==0.4.0
Jinja2==2.11.3
kaggle==1.5.10
mysql-connector-python==8.0.23
mysqlclient==2.0.3
